Restructure module as a package to be merged into the groupcache module.
