--- .github/workflow/lint.yaml
@@ -1,31 +1,0 @@
-name: Lint
-
-on:
-  push:
-    branches: [ main ]
-
-env:
-  GOLANGCI_LINT_VERSION: v1.62.2
-
-jobs:
-  lint:
-    name: lint
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Install Go
-        uses: actions/setup-go@v5
-        with:
-          go-version-file: ./go.mod
-
-      - name: Check that 'go mod tidy' was called before commit
-        run: go mod tidy && git diff --exit-code
-
-      - name: golangci-lint
-        uses: golangci/golangci-lint-action@v6
-        with:
-          version: ${{ env.GOLANGCI_LINT_VERSION }}
-          skip-cache: true # cache/restore is done by actions/setup-go step
-          args: -v
