--- .golangci.yml
@@ -1,50 +1,0 @@
-linters:
-  # Please, do not use `enable-all`: it's deprecated and will be removed soon.
-  # Inverted configuration with `enable-all` and `disable` is not scalable during updates of golangci-lint.
-  # Full list of linters - https://golangci-lint.run/usage/linters
-  disable-all: true
-  enable:
-    - gomodguard
-    - errcheck
-    - gocritic
-    - goimports
-    - gosec
-    - gosimple
-    - govet
-    - noctx
-    - nolintlint
-    - ineffassign
-    - staticcheck
-    # - stylecheck
-    - typecheck
-    - unused
-    - gocyclo
-
-linters-settings:
-  gocritic:
-    enabled-tags:
-      - diagnostic
-      - experimental
-      - opinionated
-      - performance
-      - style
-
-  govet:
-    enable-all: true
-    disable:
-     - fieldalignment
-
-issues:
-  # Maximum issues count per one linter. Set to 0 to disable. Default is 50.
-  max-issues-per-linter: 0
-
-  # Maximum count of issues with the same text. Set to 0 to disable. Default is 3.
-  max-same-issues: 50
-
-run:
-  # include test files or not, default is true
-  tests: true
-
-  # Timeout for analysis, e.g. 30s, 5m.
-  # Default: 1m
-  timeout: 5m
